fissate relazioni e test

+++ Progetto_Capstone/src/main/java/com/progetto/service/OrdineService.java
@@ -2,11 +2,15 @@ package com.progetto.service;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.progetto.interfaces.StatoOrdine;
+import com.progetto.model.Fattura;
 import com.progetto.model.Ordine;
+import com.progetto.model.Ordine.OrdineBuilder;
 import com.progetto.payload.OrdineDTO;
 import com.progetto.repository.OrdineRepository;
 import com.progetto.security.entity.User;
@@ -21,6 +25,7 @@ public class OrdineService {
 	@Autowired OrdineRepository repo;
 	@Autowired UserRepository repoUser;
 	@Autowired UserService serviceUser;
+	@Autowired FatturaService serviceFattura;
 	
 	public List<Ordine> getAll(){
 		return repo.findAll();
@@ -47,19 +52,46 @@ public class OrdineService {
 		return o;
 	}
 	
-	public Ordine update (Long id, Ordine s) {
+	public String update (Long id, Ordine s) {
 		if(!repo.existsById(id)) {
 			throw new EntityExistsException("Ordine non esistente!!");
 		}
-		return repo.save(s);
+		Ordine o = new Ordine();
+		o.setId(id);
+		o.setData(s.getData());
+		o.setStato(s.getStato());
+		o.setPagamento(s.getPagamento());
+		o.setUser(s.getUser());
+		o.setArticoliOrdinati(s.getArticoliOrdinati());
+		
+		if(s.getStato() == StatoOrdine.CONSEGNATO) {
+			Fattura f = serviceFattura.create(id);
+			f.setOrdine(o);
+			o.setFattura(f);
+		}
+		
+		repo.save(o);
+		return "Ordine modificato con successo!";
+		
+//		if(s.getStato() == StatoOrdine.CONSEGNATO || s.getStato() == StatoOrdine.IN_TRANSITO) {
+//			return "L'ordine è stato evaso e non può essere modificato";
+//		} else {
+//			repo.save(o);
+//			return "Ordine modificato con successo!";
+//		}
 	}
 	
 	public String delete (Long id) {
 		if(!repo.existsById(id)) {
 			throw new EntityExistsException("Ordine non esistente!!");
 		}
-		repo.deleteById(id);
-		return "Ordine eliminato!!";
+		Ordine o = repo.findById(id).get();
+		if(o.getStato() == StatoOrdine.CONSEGNATO || o.getStato() == StatoOrdine.IN_TRANSITO) {
+			return "L'ordine è stato evaso e non può essere cancellato";
+		} else {
+			repo.deleteById(id);
+			return "Ordine eliminato con successo!";
+		}
 	}
 
 }
